Reorder GCC linked libraries parameters last
When linking with --as-needed, which is default on some distributions, the
linking will fail if libraries are listed before the objects that reference
them.

See: https://wiki.ubuntu.com/NattyNarwhal/ToolchainTransition
See: https://wiki.debian.org/ToolChain/DSOLinking#Only_link_with_needed_libraries
Signed-off-by: Mathieu Trudel-Lapierre <mathieu.trudel-lapierre@canonical.com>

(This is an empty commit because I pulled a different copy of it with
less changelog data, but I wanted these links and docs to show up. --pjones)