Fix another place where we should be using efi_guid_global...
when we've searched for a Boot#### entry, we try to set the variables
after that, and there's one more case where we use *guid (the result of
searching) instead of efi_guid_global, but where we're not inside an
"if (found) { ... }", so it could be indeterminate.

Also related to issue#5.

Signed-off-by: Peter Jones <pjones@redhat.com>