Add fwup_set_up_update_with_buf()
fwupd (the daemon) already has the exploded firmware in a memory blob, and so
currently writes out a temp file so it can pass a fd to libfwupdate. After all
this, libfwupdate takes that fd, writes it to yet another file and does some
dark magic to make it boot next time.

This method allows us to pass a memory blob, and means we don't need the temp
file trick anymore.