libfwup: add systemd service to clean up new installs
Right now, if you re-install your OS and reformat the ESP, we'll try to
re-use the EFI variable pointing at that ESP.  We still need to validate
that device path in libfwup, but we also need to clean up the ESP in
and the variables in some corner cases.  This does that cleanup once on
each new installation.

Signed-off-by: Peter Jones <pjones@redhat.com>