Don't call GetNextVariableName again after deleting the variable.
Ard Biesheuvel noticed that we're getting an EFI_INVALID_PARAMETER from
GetNextVariableName() on the delete_boot_variable() path, which was
surprising to us all.

What appears to be happening is that when we're in Boot Services,
deleting a variable does not ever result in it existing in the
VAR_IN_DELETED_TRANSITION state; it just deletes it outright.  So the
next time you call GetNextVariableName() it can't find the variable to
start on, and then sees that VariableName[0] != 0 and just calls that
invalid.

Note that this is really only a problem because they're not caching (and
then validating) the position; they're working it out fresh each time.
This also means a GetNextVariableName() loop is O(n^2) worst case.  Bletch.

We don't really care about any of this - we only want to delete the one
variable anyway, so there's no point in calling it another time.  So
this patch just breaks out of the loop entirely if it's found the
variable.  This also fixes a minor error in the exit paths, and makes
deleting the boot order bit happen /before/ deleting the variable
itself.  We care more about not trying to boot it next reboot than about
the storage, so we'd rather we've already removed it from the boot order
if delete_variable() fails.

Signed-off-by: Peter Jones <pjones@redhat.com>
Tested-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>