Check the image size in the capsule header
Some UEFI implementation uses the all zeros GUID for the firmware class.
If the update file starts with an EFI firmware volume header, fwupd*.efi
would mistakenly treat the header as a valid capsule header. This commit
adds the conditional check for the cap image size to validate the
capsule header.

Signed-off-by: Gary Lin <glin@suse.com>